:Prettier

+++ tests/mmr/MMR.sol
@@ -167,10 +167,11 @@ contract MMR {
      *      it computes hashes recursively downward.
      *      Only appending an item calls this function
      */
-    function _getOrCreateNode(uint256 index, bytes32[] memory hashes, uint256 offset)
-        private
-        returns (bytes32)
-    {
+    function _getOrCreateNode(
+        uint256 index,
+        bytes32[] memory hashes,
+        uint256 offset
+    ) private returns (bytes32) {
         require(index <= getSize(index), "Out of range");
         if (hashes[index - offset] == bytes32(0)) {
             (uint256 leftIndex, uint256 rightIndex) = getChildren(index);
@@ -188,45 +189,45 @@ contract MMR {
     /**
      * @dev It returns a merkle proof for a leaf. Note that the index starts from 1
      */
-     function getMerkleProof(
-         bytes32[] memory hashes,
-         uint256 width,
-         uint256 index
-     ) public pure returns (bytes32[] memory, bytes32[] memory) {
-         require(index < hashes.length, "Out of range");
-         require(isLeaf(index), "Not a leaf");
+    function getMerkleProof(
+        bytes32[] memory hashes,
+        uint256 width,
+        uint256 index
+    ) public pure returns (bytes32[] memory, bytes32[] memory) {
+        require(index < hashes.length, "Out of range");
+        require(isLeaf(index), "Not a leaf");
 
-         // Find all peaks for bagging
-         uint256[] memory peaks = getPeakIndexes(width);
+        // Find all peaks for bagging
+        uint256[] memory peaks = getPeakIndexes(width);
 
-         bytes32[] memory peakBaggingArray = new bytes32[](peaks.length);
-         uint256 cursor;
-         for (uint256 i = 0; i < peaks.length; i++) {
-             // Collect the hash of all peaks
-             peakBaggingArray[i] = hashes[peaks[i]];
-             // Find the peak which includes the target index
-             if (peaks[i] >= index && cursor == 0) {
-                 cursor = peaks[i];
-             }
-         }
-         uint256 left;
-         uint256 right;
+        bytes32[] memory peakBaggingArray = new bytes32[](peaks.length);
+        uint256 cursor;
+        for (uint256 i = 0; i < peaks.length; i++) {
+            // Collect the hash of all peaks
+            peakBaggingArray[i] = hashes[peaks[i]];
+            // Find the peak which includes the target index
+            if (peaks[i] >= index && cursor == 0) {
+                cursor = peaks[i];
+            }
+        }
+        uint256 left;
+        uint256 right;
 
-         // Get hashes of the siblings in the mountain which the index belongs to.
-         // It moves the cursor from the summit of the mountain down to the target index
-         uint8 height = heightAt(cursor);
-         bytes32[] memory siblings = new bytes32[](height - 1);
-         while (cursor != index) {
-             height--;
-             (left, right) = getChildren(cursor);
-             // Move the cursor down to the left side or right side
-             cursor = index <= left ? left : right;
-             // Remaining node is the sibling
-             siblings[height - 1] = hashes[index <= left ? right : left];
-         }
+        // Get hashes of the siblings in the mountain which the index belongs to.
+        // It moves the cursor from the summit of the mountain down to the target index
+        uint8 height = heightAt(cursor);
+        bytes32[] memory siblings = new bytes32[](height - 1);
+        while (cursor != index) {
+            height--;
+            (left, right) = getChildren(cursor);
+            // Move the cursor down to the left side or right side
+            cursor = index <= left ? left : right;
+            // Remaining node is the sibling
+            siblings[height - 1] = hashes[index <= left ? right : left];
+        }
 
-         return (peakBaggingArray, siblings);
-     }
+        return (peakBaggingArray, siblings);
+    }
 
     // Returns the closest power of two for a number
     function closestPow2(uint256 number) public pure returns (uint256) {
@@ -254,7 +255,11 @@ contract MMR {
             // Put the hashed leaf to the array
             hashes[getSize(index) + 1] = leaf;
         }
-        bytes32 subpeak = _getOrCreateNode(getSize(subpeakWidth + offset), hashes, getSize(offset));
+        bytes32 subpeak = _getOrCreateNode(
+            getSize(subpeakWidth + offset),
+            hashes,
+            getSize(offset)
+        );
         return (subpeak, offset + subpeakWidth);
     }
 
@@ -266,7 +271,7 @@ contract MMR {
         bytes32 subpeak;
         uint256 subpeaksNumber = numOfPeaks(data.length);
         bytes32[] memory subpeaks = new bytes32[](subpeaksNumber);
-        for (uint i = 0; i < subpeaksNumber; i++) {
+        for (uint256 i = 0; i < subpeaksNumber; i++) {
             (subpeak, offset) = getSubpeak(data, offset);
             subpeaks[i] = subpeak;
         }
